refactor: ReservationTime Controller DTO 네이밍 변경
