+++ BE/src/main/java/roomescape/reservationTime/presentation/dto/response/ReservationTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.reservationTime.presentation.dto;
+package roomescape.reservationTime.presentation.dto.response;
 
 import roomescape.reservationTime.domain.ReservationTime;
